+++ sdk/cosmos/azure_data_cosmos/src/clients/mod.rs
@@ -7,7 +7,8 @@ mod container_client;
 mod cosmos_client;
 mod cosmos_client_builder;
 mod database_client;
-mod offers_client;
+pub(crate) mod offers_client;
+mod throughput_poller;
 
 use std::sync::Arc;
 
@@ -38,4 +39,4 @@ pub use container_client::ContainerClient;
 pub use cosmos_client::CosmosClient;
 pub use cosmos_client_builder::CosmosClientBuilder;
 pub use database_client::DatabaseClient;
-pub(crate) use offers_client::OffersClient;
+pub use throughput_poller::ThroughputPoller;