+++ sdk/cosmos/azure_data_cosmos/tests/split_tests/mod.rs
@@ -1,0 +1,6 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+mod cosmos_split_offers;
+
+#[path = "../framework/mod.rs"]
+mod framework;